Refactor data_queues to be more pluggable
