+++ geonode/geonodegp/celery_app.py
@@ -9,21 +9,6 @@ os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'geonodegp.settings')
 
 from django.conf import settings
 
-from celery.schedules import crontab
-settings.CELERYBEAT_SCHEDULE = {
-    'gfms': {
-        'task': 'geonodegp.data_queues.tasks.gfms_task',
-        #'schedule': crontab(hour='*/3', minute=30),
-        'schedule': crontab(minute='*/2'),
-        'args': ()
-    },
-    'forecast_io': {
-        'task': 'geonodegp.data_queues.tasks.forecast_io_task',
-        'schedule': crontab(minute=2),
-        'args': ()
-    },
-}
-
 app = Celery('geonodegp')
 
 # Using a string here means the worker will not have to
