+++ geonode/geonodegp/data_queues/tasks.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import
 
 from celery import shared_task
-from .processors import *
+from .processor_base import *
 
 
 @shared_task()
